+++ spinetoolbox/widgets/settings_widget.py
@@ -18,8 +18,9 @@ Widget for controlling user settings.
 
 import logging
 import os
-from PySide2.QtWidgets import QWidget, QStatusBar, QFileDialog, QStyle
+from PySide2.QtWidgets import QWidget, QStatusBar, QFileDialog, QStyle, QColorDialog
 from PySide2.QtCore import Slot, Qt
+from PySide2.QtGui import QPixmap, QColor
 import ui.settings
 from config import DEFAULT_PROJECT_DIR, DEFAULT_WORK_DIR, STATUSBAR_SS, \
     SETTINGS_SS, GAMS_EXECUTABLE, GAMSIDE_EXECUTABLE, JULIA_EXECUTABLE
@@ -42,6 +43,8 @@ class SettingsWidget(QWidget):
         self._project = self._toolbox.project()
         self._qsettings = self._toolbox.qsettings()
         self.orig_work_dir = ""  # Work dir when this widget was opened
+        # Initial scene bg color. Is overridden immediately in read_settings() if it exists in qSettings
+        self.bg_color = self._toolbox.ui.graphicsView.scene().bg_color
         # Set up the ui from Qt Designer files
         self.ui = ui.settings.Ui_SettingsForm()
         self.ui.setupUi(self)
@@ -72,6 +75,9 @@ class SettingsWidget(QWidget):
         self.ui.toolButton_browse_gams.clicked.connect(self.browse_gams_path)
         self.ui.toolButton_browse_julia.clicked.connect(self.browse_julia_path)
         self.ui.toolButton_browse_work.clicked.connect(self.browse_work_path)
+        self.ui.toolButton_bg_color.clicked.connect(self.show_color_dialog)
+        self.ui.radioButton_bg_grid.clicked.connect(self.update_scene_bg)
+        self.ui.radioButton_bg_solid.clicked.connect(self.update_scene_bg)
 
     @Slot(bool, name="browse_gams_path")
     def browse_gams_path(self, checked=False):
@@ -121,6 +127,44 @@ class SettingsWidget(QWidget):
         selected_path = os.path.abspath(answer)
         self.ui.lineEdit_work_dir.setText(selected_path)
 
+    @Slot(bool, name="show_color_dialog")
+    def show_color_dialog(self, checked=False):
+        """Let user pick the bg color.
+
+        Args:
+            checked (boolean): Value emitted with clicked signal
+        """
+        # noinspection PyArgumentList
+        color = QColorDialog.getColor(initial=self.bg_color)
+        if not color.isValid():
+            return  # Canceled
+        self.ui.radioButton_bg_solid.setChecked(True)
+        self.bg_color = color
+        self.update_bg_color()
+
+    def update_bg_color(self):
+        """Set tool button icon as the selected color and update
+        Design View scene background color."""
+        pixmap = QPixmap(16, 16)
+        pixmap.fill(self.bg_color)
+        self.ui.toolButton_bg_color.setIcon(pixmap)
+        self._toolbox.ui.graphicsView.scene().set_bg_color(self.bg_color)
+        self._toolbox.ui.graphicsView.scene().update()
+
+    @Slot(bool, name="update_scene_bg")
+    def update_scene_bg(self, checked):
+        """Set background drawing type on scene depending on radiobutton states.
+
+        Args:
+            checked (boolean): Toggle state
+        """
+        if self.ui.radioButton_bg_grid.isChecked():
+            self._toolbox.ui.graphicsView.scene().set_bg_grid(True)
+            self._toolbox.ui.graphicsView.scene().update()
+        elif self.ui.radioButton_bg_solid.isChecked():
+            self._toolbox.ui.graphicsView.scene().set_bg_grid(False)
+            self._toolbox.ui.graphicsView.scene().update()
+
     def read_settings(self):
         """Read current settings from config object and update UI to show them."""
         open_previous_project = self._configs.getboolean("settings", "open_previous_project")
@@ -135,6 +179,8 @@ class SettingsWidget(QWidget):
         use_repl = self._configs.getboolean("settings", "use_repl")
         julia_path = self._configs.get("settings", "julia_path")
         delete_data = self._configs.getboolean("settings", "delete_data")
+        bg_grid = self._qsettings.value("appSettings/bgGrid", defaultValue="false")
+        bg_color = self._qsettings.value("appSettings/bgColor", defaultValue="false")
         if open_previous_project:
             self.ui.checkBox_open_previous_project.setCheckState(Qt.Checked)
         if show_exit_prompt:
@@ -157,6 +203,15 @@ class SettingsWidget(QWidget):
             self.ui.checkBox_commit_at_exit.setCheckState(Qt.PartiallyChecked)
         if smooth_zoom == "true":
             self.ui.checkBox_use_smooth_zoom.setCheckState(Qt.Checked)
+        if bg_grid == "true":
+            self.ui.radioButton_bg_grid.setChecked(True)
+        else:
+            self.ui.radioButton_bg_solid.setChecked(True)
+        if bg_color == "false":
+            pass
+        else:
+            self.bg_color = bg_color
+        self.update_bg_color()
         if datetime:
             self.ui.checkBox_datetime.setCheckState(Qt.Checked)
         if delete_data:
@@ -188,6 +243,7 @@ class SettingsWidget(QWidget):
         g = str(int(self.ui.checkBox_commit_at_exit.checkState()))
         h = True if int(self.ui.checkBox_use_smooth_zoom.checkState()) else False
         d = int(self.ui.checkBox_datetime.checkState())
+        bg_grid = True if self.ui.radioButton_bg_grid.isChecked() else False
         delete_data = int(self.ui.checkBox_delete_data.checkState())
         # Check that GAMS directory is valid. Set it empty if not.
         gams_path = self.ui.lineEdit_gams_path.text()
@@ -216,6 +272,8 @@ class SettingsWidget(QWidget):
         self._configs.set("settings", "gams_path", gams_path)
         self._configs.setboolean("settings", "use_repl", e)
         self._configs.set("settings", "julia_path", julia_path)
+        self._qsettings.setValue("appSettings/bgGrid", bg_grid)
+        self._qsettings.setValue("appSettings/bgColor", self.bg_color)
         # Update project settings
         self.update_project_settings()
         self._configs.save()
@@ -292,8 +350,6 @@ class SettingsWidget(QWidget):
         Args:
             e (QMouseEvent): Mouse event
         """
-        # logging.debug("MouseMoveEvent at pos:%s" % e.pos())
-        # logging.debug("MouseMoveEvent globalpos:%s" % e.globalPos())
         currentpos = self.pos()
         globalpos = e.globalPos()
         if not self._mouseMovePos: